\name{CESens-package}
\alias{CESens-package}
%\alias{CESens}
\docType{package}
\title{
Sensitivity Analysis For Causal Effect Estimation
}
\description{
An interactive tool for estimating principal stratification-based causal effects 
  complete with sensitivity analysis.  User can define distributions for factual and counterfactual
  trial data and compute causal effect and 95\% confidence intervals.
  
  Currently this package is only supported in the Windows environment.  Testing has been limited to
  Windows 7.
}
\details{
\tabular{ll}{
Package: \tab CESens\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-14\cr
License: \tab GPL-2\cr
}
Reference help file for CESens().
}
\author{
Lisa Henn

Maintainer: Julian Wolfson <julianw@umn.edu>
}
\references{
Bowman, Adrian et al., \dQuote{rpanel:  Simple Interactive Controls 
  for R Functions Using the tcltk Package}, \emph{Journal of Statistical 
  Software}, Jan 2007, Vol 17, Issue 9.
 See also \url{http://www.stats.gla.ac.uk/~adrian/rpanel/}.
}

\keyword{ package }
\keyword{ model }
\keyword{ nonparametric }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
